requirements.txt
# Core frameworks
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langgraph>=0.3.0

# Web framework
flask==3.1.0

# Database connectors
sqlalchemy==2.0.35
pymysql==1.1.1  # For MySQL
psycopg2-binary==2.9.10  # For PostgreSQL

# Utilities
python-dotenv==1.0.1
pandas==2.2.3
tabulate==0.9.0  # For better table formatting
pyyaml>=6.0  # For domain configuration